Fix the Barrier powerup exploding
Upgrades::Adrenaline: Fix logs also showing up in the full log
Player: Change discount to a float instead of a double
Common: Fix IsSmallNumber not even working

+++ pk7/lcode/Main/inc/lith_player.h
@@ -97,7 +97,7 @@ typedef struct player_s
    int maxarmor;
    fixed jumpheight;
    fixed viewheight;
-   double discount;
+   float discount;
    
    // Additive view
    float addpitch;
